libstdc++: Write timestamp to libstdc++-performance.sum file
The results of 'make check-performance' are appended to the .sum file,
with no indication where one set of results ends and the next begins. We
could just remove the file when starting a new run, but appending makes
it a little easier to compare with previous runs, without having to copy
and store old files.

This adds a header containing a timestamp to the file when starting a
new run.

libstdc++-v3/ChangeLog:

	* scripts/check_performance: Add timestamp to output file at
	start of run.