ipa: Introduce a one jump function dumping function
I plan to introduce a verifier that prints a single jump function when
it fails with the function introduced in this one.  Because it is a
verifier, the risk that it would need to e reverted are non-zero and
because the function can be useful on its own, this is a special patch
to introduce it.

gcc/ChangeLog:

2024-11-01  Martin Jambor  <mjambor@suse.cz>

	* ipa-prop.h (ipa_dump_jump_function): Declare.
	* ipa-prop.cc (ipa_dump_jump_function): New function.
	(ipa_print_node_jump_functions_for_edge): Move printing of
	individual jump functions to the new function.