Fix various sh tests to work with c23
A few SH tests want to create a bool typedef which doesn't work for c23.
Easiest fix which should have no impact on the test behavior would be to just
change the name of the typedef so that doesn't conflict.

One test has a crazy function signature (similar to the PRU test someone just
fixed up).  For that I'm using -std=gnu17.

Pushing to the trunk.

testsuite/
	* gcc.target/sh/pr51244-15.c: Use "mybool" rather than "bool".
	* gcc.target/sh/pr52933-1.c: Similarly.
	* gcc.target/sh/pr54089-1.c: Similarly.
	* gcc.target/sh/pr54089-7.c: Similarly.
	* gcc.target/sh/pr54089-8.c: Similarly.
	* gcc.target/sh/pr54089-9.c: Similarly.
	* gcc.target/sh/pr64366.c: Use -std=gnu17.