ada: Add minimal support for other delayed aspects on controlled objects
This extends the processing done for the Address aspect to other delayed
aspects.  The External_Name aspect is also reclassified as a representation
aspect and the three representation aspects External_Name, Link_Name and
Linker_Section are moved from the Always_Delay to the Rep_Aspect category,
which makes it possible not to delay them in most cases with a small tweak.

gcc/ada/ChangeLog:

	* aspects.ads (Is_Representation_Aspect): True for External_Name.
	(Aspect_Delay): Use Rep_Aspect for External_Name, Link_Name and
	Linker_Section.
	* einfo.ads (Initialization_Statements): Document extended usage.
	* exp_util.adb (Needs_Initialization_Statements): Return True for
	all delayed aspects.
	* freeze.adb (Check_Address_Clause): Do not move the initialization
	expression here...
	(Freeze_Object_Declaration): ...but here instead, as well as for all
	delayed aspects.  Remove test for pragma Linker_Section.
	* sem_ch13.adb (Analyze_One_Aspect): Do not delay in the Rep_Aspect
	case if the expression is a string literal.