testsuite: Fix pr101145inf*.c testcases [PR117494]
Instead of doing a dg-run with a specific target check for linux.
Use signal as the effective-target since this requires the use
of ALARM signal to do the testing.
Also use check_vect in the main and renames main to main1 to make sure
we don't use the registers.

Tested on x86_64-linux-gnu.

	PR testsuite/117494
gcc/testsuite/ChangeLog:

	* gcc.dg/vect/pr101145inf.c: Remove dg-do and replace
	with dg-require-effective-target of signal.
	* gcc.dg/vect/pr101145inf_1.c: Likewise.
	* gcc.dg/vect/pr101145inf.inc: Rename main to main1
	and mark as noinline.
	Include tree-vect.h. Have main call check_vect and main1.

Signed-off-by: Andrew Pinski <quic_apinski@quicinc.com>