c++: reduce redundant deprecated warnings
If a template uses a deprecated function, we should warn there and not also
whenever the template is instantiated.  I implement this by suppressing
the warning at the location; then to make this also work with modules, I
need to make sure to set TREE_NO_WARNING so that the warning spec for this
location gets recorded.

And then I noticed that has_warning_spec was broken such that if it
returned true than get_nowarn_spec would always return null.

gcc/cp/ChangeLog:

	* decl2.cc (cp_handle_deprecated_or_unavailable): Avoid redundant
	warning.
	* call.cc (build_over_call): Set TREE_NO_WARNING for calls
	to deprecated functions.
	* semantics.cc (finish_call_expr): Propagate TREE_NO_WARNING.

gcc/ChangeLog:

	* warning-control.cc (has_warning_spec): Fix handling of
	get_no_warning_bit.

gcc/testsuite/ChangeLog:

	* g++.dg/warn/deprecated-21.C: New test.
	* g++.dg/modules/warn-spec-2_a.C: New test.
	* g++.dg/modules/warn-spec-2_b.C: New test.