aarch64: Add early_ra::record_live_range_failure
So far, early_ra has used a single m_allocation_successful bool
to record whether the current region is still being allocated.
But there are (at least) two reasons why we might pull out of
attempting an allocation:

(1) We can't track the liveness of individual FP allocnos,
    due to some awkward subregs.

(2) We're afraid of doing a worse job than the proper allocator.

A later patch needs to distinguish (1) from other reasons, since
(1) means that the liveness information is not trustworthy.
(Currently we assume it is not trustworthy whenever
m_allocation_successful is false, but that's too conservative.)

gcc/
	* config/aarch64/aarch64-early-ra.cc
	(early_ra::record_live_range_failure): New member function.
	(early_ra::m_accurate_live_ranges): New member variable.
	(early_ra::start_new_region): Set m_accurate_live_ranges to true.
	(early_ra::get_allocno_subgroup): Use record_live_range_failure
	to abort the allocation on invalid subregs.