[PATCH] testsuite: Require C99 for pow-to-ldexp.c
pow-to-ldexp.c checks for calls to __builtin_ldexpf and __builtin_ldexpl, which
will only be performed when the compiler knows the target has a C99 libm
available.

Modified the test to add a C99 runtime requirement.

This fixes the failure on arm-eabi targets for this test case.

Signed-off-by: Soumya AR <soumyaa@nvidia.com>

gcc/testsuite/ChangeLog:

	* gcc.dg/tree-ssa/pow-to-ldexp.c: Require c99_runtime.