ada: Accept static strings with External_Initialization
Before this patch, the argument to the External_Initialization aspect
had to be a string literal. This patch extends the possibilities so that
any static string is accepted.

A new helper function, Is_OK_Static_Expression_Of_Type, is introduced,
and in addition to the main change of this patch a couple of calls to
that helper function are added in other places to replace equivalent
inline code.

gcc/ada/ChangeLog:

	* sem_eval.ads (Is_OK_Static_Expression_Of_Type): New function.
	* sem_eval.adb (Is_OK_Static_Expression_Of_Type): Likewise.
	* sem_ch13.adb (Check_Expr_Is_OK_Static_Expression): Use new function.
	* sem_prag.adb (Check_Expr_Is_OK_Static_Expression): Likewise.
	* sem_ch3.adb (Apply_External_Initialization): Accept static strings
	for the parameter.